requirements.txt
Django==4.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
django-crontab==0.7.1
python-dotenv~=1.0.1
django-apscheduler==0.6.2
python-dateutil==2.9.0
django-crispy-forms==2.2
crispy-bootstrap4==2024.1
ipython==8.25.0
redis==5.0.6